requirements.txt
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
Authlib==1.6.0
aws-requests-auth==0.4.3
backoff==2.2.1
beautifulsoup4==4.13.4
boto3==1.38.40
botocore==1.38.40
bs4==0.0.2
cachetools==5.5.2
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
cryptography==45.0.5
Deprecated==1.2.18
dill==0.4.0
distro==1.9.0
dnspython==2.7.0
docstring_parser==0.16
email_validator==2.2.0
Events==0.5
exceptiongroup==1.3.0
fastapi==0.115.13
fastmcp==2.10.2
filelock==3.18.0
fsspec==2025.5.1
google-api-core==2.25.1
google-api-python-client==2.173.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
googleapis-common-protos==1.70.0
grpcio==1.73.0
h11==0.16.0
h2==4.2.0
halo==0.0.31
hf-xet==1.1.2
hpack==4.1.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.32.3
hyperframe==6.1.0
idna==3.10
importlib_metadata==8.7.0
jiter==0.10.0
jmespath==1.0.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
log-symbols==0.0.14
markdown-it-py==3.0.0
mcp==1.10.1
mdurl==0.1.2
mem0ai==0.1.109
mpmath==1.3.0
numpy==2.3.0
ollama==0.5.1
openai==1.88.0
openapi-pydantic==0.5.1
opensearch-py==2.8.0
opentelemetry-api==1.34.1
opentelemetry-exporter-otlp-proto-common==1.33.1
opentelemetry-exporter-otlp-proto-http>=1.33.1,<2.0.0
opentelemetry-proto==1.33.1
opentelemetry-sdk>=1.33.1,<2.0.0
opentelemetry-semantic-conventions==0.55b1
packaging==24.2
pillow==11.2.1
pinecone==7.0.2
pinecone-plugin-assistant==1.6.1
pinecone-plugin-interface==0.0.7
portalocker==2.10.1
posthog==5.3.0
prompt_toolkit==3.0.51
proto-plus==1.26.1
protobuf==6.31.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.9.1
pydantic_core==2.33.2
Pygments==2.19.1
PyJWT==2.10.1
PyMuPDF==1.26.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
qdrant-client==1.14.3
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
rich==14.0.0
rpds-py==0.26.0
rsa==4.9.1
s3transfer==0.13.0
safetensors==0.5.3
shellingham==1.5.4
six==1.17.0
slack_bolt==1.23.0
slack_sdk==3.35.0
sniffio==1.3.1
soupsieve==2.7
spinners==0.0.24
SQLAlchemy==2.0.41
sse-starlette==2.3.6
starlette==0.46.2
strands-agents==0.1.8
strands-agents-builder==0.1.3
strands-agents-tools==0.1.6
sympy==1.14.0
tenacity==9.1.2
termcolor==3.1.0
tokenizers==0.21.1
tqdm==4.67.1
transformers==4.52.4
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.0
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.34.3
watchdog==6.0.0
wcwidth==0.2.13
wrapt==1.17.2
zipp==3.23.0
